Move seed file
