+++ src/backend/db/lib/index.ts
@@ -1,4 +1,4 @@
 export * from "./get-db";
 export * from "./migrate";
 export * from "./reset";
-export * from "./seed";
+export * from "../../tests/seed";
